# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StageLink API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /artists:
    get:
      operationId: getArtists
      tags: [artists]
      summary: List/discover artists with filters
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: genre
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: minBudget
          in: query
          schema:
            type: integer
        - name: maxBudget
          in: query
          schema:
            type: integer
        - name: minRating
          in: query
          schema:
            type: number
        - name: verified
          in: query
          schema:
            type: boolean
        - name: artistType
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated list of artists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistListResponse"
...
  /artists/{artistId}:
    get:
      operationId: getArtist
      tags: [artists]
      summary: Get full artist profile
      parameters:
        - name: artistId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Full artist profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistProfile"
        "404":
          description: Not found
  /artists/me:
    get:
      operationId: getMyArtistProfile
      tags: [artists]
      summary: Get own artist profile
      responses:
        "200":
          description: Own artist profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistProfile"
    put:
      operationId: updateArtistProfile
      tags: [artists]
      summary: Update own artist profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtistProfileInput"
      responses:
        "200":
          description: Updated artist profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistProfile"
  /artists/{artistId}/reviews:
    get:
      operationId: getArtistReviews
      tags: [artists]
      summary: Get reviews for an artist
      parameters:
        - name: artistId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [artists]
      summary: Submit a review for an artist
      parameters:
        - name: artistId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ── Tracks ─────────────────────────────────────────────────────────────────
...
  /bookings:
    get:
      operationId: getBookings
      tags: [bookings]
      summary: Get current user's bookings
      responses:
        "200":
          description: Bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Request a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{bookingId}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking details
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{bookingId}/status:
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status (artist accepts/declines)
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingStatusUpdate"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # ── Posts ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ArtistCard:
      type: object
      required: [id, displayName, artistType, genres, country, coverImageUrl, rating, followersCount, verified, xp, level]
      properties:
        id:
          type: integer
        displayName:
          type: string
        artistType:
          type: string
        genres:
          type: array
          items:
            type: string
        country:
          type: string
        city:
          type: ["string", "null"]
        coverImageUrl:
          type: string
        avatarUrl:
          type: ["string", "null"]
        rating:
          type: number
        reviewCount:
          type: integer
        followersCount:
          type: integer
        verified:
          type: boolean
        xp:
          type: integer
        level:
          type: integer
        bookingPrice:
          type: ["integer", "null"]
    ArtistProfile:
      type: object
      required: [id, displayName, artistType, genres, country, bio, coverImageUrl, rating, followersCount, verified, xp, level, socialLinks, tracks, isFollowedByMe]
      properties:
        id:
          type: integer
        displayName:
          type: string
        artistType:
          type: string
        genres:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            type: string
        country:
          type: string
        city:
          type: ["string", "null"]
        bio:
          type: string
        coverImageUrl:
          type: string
        avatarUrl:
          type: ["string", "null"]
        rating:
          type: number
        reviewCount:
          type: integer
        followersCount:
          type: integer
        verified:
          type: boolean
        xp:
          type: integer
        level:
          type: integer
        badges:
          type: array
          items:
            type: string
        bookingPrice:
          type: ["integer", "null"]
        socialLinks:
          $ref: "#/components/schemas/SocialLinks"
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/Track"
        isFollowedByMe:
          type: boolean
    ArtistProfileInput:
      type: object
      properties:
        displayName:
          type: string
        artistType:
          type: string
        genres:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            type: string
        country:
          type: string
        city:
          type: string
        bio:
          type: string
        coverImageUrl:
          type: string
        avatarUrl:
          type: string
        bookingPrice:
          type: integer
        socialLinks:
          $ref: "#/components/schemas/SocialLinks"
    SocialLinks:
      type: object
      properties:
        instagram:
          type: ["string", "null"]
        twitter:
          type: ["string", "null"]
        youtube:
          type: ["string", "null"]
        spotify:
          type: ["string", "null"]
        soundcloud:
          type: ["string", "null"]
        tiktok:
          type: ["string", "null"]
    ArtistListResponse:
      type: object
      required: [artists, total, page, limit]
      properties:
        artists:
          type: array
          items:
            $ref: "#/components/schemas/ArtistCard"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    Review:
      type: object
      required: [id, artistId, reviewerId, reviewerName, rating, comment, createdAt]
      properties:
        id:
          type: integer
        artistId:
          type: integer
        reviewerId:
          type: integer
        reviewerName:
          type: string
        reviewerAvatarUrl:
          type: ["string", "null"]
        rating:
          type: number
        comment:
          type: string
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [rating, comment]
      properties:
        rating:
          type: number
          minimum: 1
          maximum: 5
        comment:
          type: string
          minLength: 10

    # ── Tracks ────────────────────────────────────────────────────────────
    Track:
      type: object
      required: [id, artistId, title, genre, trackType, durationSeconds, coverUrl, plays, createdAt]
      properties:
        id:
          type: integer
        artistId:
          type: integer
        artistName:
          type: string
        title:
          type: string
        genre:
          type: string
        trackType:
          type: string
          enum: [single, album, beat, instrumental, sample]
        durationSeconds:
          type: integer
        coverUrl:
          type: string
        audioUrl:
          type: ["string", "null"]
        price:
          type: ["integer", "null"]
        plays:
          type: integer
        likes:
          type: integer
        createdAt:
          type: string
...
    Booking:
      type: object
      required: [id, artistId, clientId, eventType, eventDate, location, budget, status, message, createdAt]
      properties:
        id:
          type: integer
        artistId:
          type: integer
        artistName:
          type: string
        artistAvatarUrl:
          type: ["string", "null"]
        clientId:
          type: integer
        clientName:
          type: string
        eventType:
          type: string
        eventDate:
          type: string
        location:
          type: string
        budget:
          type: integer
        status:
          type: string
          enum: [pending, accepted, declined, completed, cancelled]
        message:
          type: string
        createdAt:
          type: string
    BookingInput:
      type: object
      required: [artistId, eventType, eventDate, location, budget, message]
      properties:
        artistId:
          type: integer
        eventType:
          type: string
        eventDate:
          type: string
        location:
          type: string
        budget:
          type: integer
        message:
          type: string
    BookingStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, declined, completed, cancelled]

    # ── Posts ─────────────────────────────────────────────────────────────
...