# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * StageLink API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  artistId: number;
  artistName?: string;
  /** @nullable */
  artistAvatarUrl?: string | null;
  clientId: number;
  clientName?: string;
  eventType: string;
  eventDate: string;
  location: string;
  budget: number;
  status: BookingStatus;
  message: string;
  createdAt: string;
}