# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * StageLink API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  artistId: number;
  eventType: string;
  eventDate: string;
  location: string;
  budget: number;
  message: string;
}