# lib/api-zod/src/generated/types/bookingStatusUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * StageLink API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatusUpdateStatus } from './bookingStatusUpdateStatus';

export interface BookingStatusUpdate {
  status: BookingStatusUpdateStatus;
}